requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
streamlit==1.28.2
face_recognition==1.3.0
faiss-cpu>=1.9.0.post1
numpy<2.0.0
requests==2.31.0
python-multipart==0.0.6
websockets==12.0
opencv-python==4.8.1.78
cryptography
dlib